Centralize template rendering
