+++ app/routes/device_types.py
@@ -1,13 +1,13 @@
 from fastapi import APIRouter, Request, Depends, HTTPException, Form
 from fastapi.responses import RedirectResponse
-from fastapi.templating import Jinja2Templates
+from app.utils.templates import templates
 from sqlalchemy.orm import Session
 
 from app.utils.db_session import get_db
 from app.utils.auth import require_role
 from app.models.models import DeviceType
 
-templates = Jinja2Templates(directory="app/templates")
+
 
 router = APIRouter()
