+++ app/routes/task_views.py
@@ -1,12 +1,12 @@
 from fastapi import APIRouter, Request, Depends, HTTPException
-from fastapi.templating import Jinja2Templates
+from app.utils.templates import templates
 from sqlalchemy.orm import Session
 
 from app.utils.db_session import get_db
 from app.utils.auth import get_current_user
 from app.models.models import ConfigBackup
 
-templates = Jinja2Templates(directory="app/templates")
+
 
 router = APIRouter()
