+++ app/routes/welcome.py
@@ -1,10 +1,10 @@
 from fastapi import APIRouter, Request, Depends
-from fastapi.templating import Jinja2Templates
+from app.utils.templates import templates
 
 from app.utils.auth import get_current_user
 
 router = APIRouter()
-templates = Jinja2Templates(directory="app/templates")
+
 
 WELCOME_TEXT = {
     "viewer": [
